fix: double to float

+++ db/migrations/init.sql
@@ -4,7 +4,7 @@ CREATE TABLE `profile` (
     `initial_login_date` TIMESTAMP NOT NULL,
     `last_login_date` TIMESTAMP,
     `play_time` BIGINT TIMESTAMP 0,
-    `experiences` DOUBLE DEFAULT 0.0, 
+    `experiences` FLOAT DEFAULT 0.0, 
     `currency` INT DEFAULT 0, 
     `total_build_blocks` INT DEFAULT 0,
     `total_destroy_blocks` INT DEFAULT 0,
